Improve challenge time processing
[-] Store the save container in the state

[-] Auto-import existing profiles and times when a new save file is created and load best times from .rec files on the save loading

[-] Use best times from the generic save
